Move config specific helpers to new file
Config specific helpers and associated types moved to new
`configuration.mjs` file. Move the associated tests to new
`configuration` folder within `common`, remove configuration
helper tests from `common/index.jsdom.test.mjs`. Update component
references to the functions from `configuration.mjs` and update the
documentation.
