+++ packages/govuk-frontend/src/govuk/common/index.mjs
@@ -1,5 +1,3 @@
-import { normaliseString } from './normalise-string.mjs'
-
 /**
  * Common helpers which do not require polyfill.
  *
@@ -7,101 +5,6 @@ import { normaliseString } from './normalise-string.mjs'
  * so that the polyfill can be properly tree-shaken and does not burden
  * the components that do not need that helper
  */
-
-/**
- * Config merging function
- *
- * Takes any number of objects and combines them together, with
- * greatest priority on the LAST item passed in.
- *
- * @internal
- * @param {...{ [key: string]: unknown }} configObjects - Config objects to merge
- * @returns {{ [key: string]: unknown }} A merged config object
- */
-export function mergeConfigs(...configObjects) {
-  // Start with an empty object as our base
-  /** @type {{ [key: string]: unknown }} */
-  const formattedConfigObject = {}
-
-  // Loop through each of the passed objects
-  for (const configObject of configObjects) {
-    for (const key of Object.keys(configObject)) {
-      const option = formattedConfigObject[key]
-      const override = configObject[key]
-
-      // Push their keys one-by-one into formattedConfigObject. Any duplicate
-      // keys with object values will be merged, otherwise the new value will
-      // override the existing value.
-      if (isObject(option) && isObject(override)) {
-        // @ts-expect-error Index signature for type 'string' is missing
-        formattedConfigObject[key] = mergeConfigs(option, override)
-      } else {
-        // Apply override
-        formattedConfigObject[key] = override
-      }
-    }
-  }
-
-  return formattedConfigObject
-}
-
-/**
- * Extracts keys starting with a particular namespace from dataset ('data-*')
- * object, removing the namespace in the process, normalising all values
- *
- * @internal
- * @param {{ schema: Schema }} Component - Component class
- * @param {DOMStringMap} dataset - The object to extract key-value pairs from
- * @param {string} namespace - The namespace to filter keys with
- * @returns {ObjectNested | undefined} Nested object with dot-separated key namespace removed
- */
-export function extractConfigByNamespace(Component, dataset, namespace) {
-  const property = Component.schema.properties[namespace]
-
-  // Only extract configs for object schema properties
-  if (property?.type !== 'object') {
-    return
-  }
-
-  // Add default empty config
-  const newObject = {
-    [namespace]: /** @type {ObjectNested} */ ({})
-  }
-
-  for (const [key, value] of Object.entries(dataset)) {
-    /** @type {ObjectNested | ObjectNested[NestedKey]} */
-    let current = newObject
-
-    // Split the key into parts, using . as our namespace separator
-    const keyParts = key.split('.')
-
-    /**
-     * Create new level per part
-     *
-     * e.g. 'i18n.textareaDescription.other' becomes
-     * `{ i18n: { textareaDescription: { other } } }`
-     */
-    for (const [index, name] of keyParts.entries()) {
-      if (typeof current === 'object') {
-        // Drop down to nested object until the last part
-        if (index < keyParts.length - 1) {
-          // New nested object (optionally) replaces existing value
-          if (!isObject(current[name])) {
-            current[name] = {}
-          }
-
-          // Drop down into new or existing nested object
-          current = current[name]
-        } else if (key !== namespace) {
-          // Normalised value (optionally) replaces existing value
-          current[name] = normaliseString(value)
-        }
-      }
-    }
-  }
-
-  return newObject[namespace]
-}
 
 /**
  * Get hash fragment from URL
@@ -221,44 +124,6 @@ export function isSupported($scope = document.body) {
 }
 
 /**
- * Validate component config by schema
- *
- * Follows limited examples in JSON schema for wider support in future
- *
- * {@link https://ajv.js.org/json-schema.html#compound-keywords}
- * {@link https://ajv.js.org/packages/ajv-errors.html#single-message}
- *
- * @internal
- * @param {Schema} schema - Config schema
- * @param {{ [key: string]: unknown }} config - Component config
- * @returns {string[]} List of validation errors
- */
-export function validateConfig(schema, config) {
-  const validationErrors = []
-
-  // Check errors for each schema
-  for (const [name, conditions] of Object.entries(schema)) {
-    const errors = []
-
-    // Check errors for each schema condition
-    if (Array.isArray(conditions)) {
-      for (const { required, errorMessage } of conditions) {
-        if (!required.every((key) => !!config[key])) {
-          errors.push(errorMessage) // Missing config key value
-        }
-      }
-
-      // Check one condition passes or add errors
-      if (name === 'anyOf' && !(conditions.length - errors.length >= 1)) {
-        validationErrors.push(...errors)
-      }
-    }
-  }
-
-  return validationErrors
-}
-
-/**
  * Check for an array
  *
  * @internal
@@ -276,7 +141,7 @@ function isArray(option) {
  * @param {unknown} option - Option to check
  * @returns {boolean} Whether the option is an object
  */
-function isObject(option) {
+export function isObject(option) {
   return !!option && typeof option === 'object' && !isArray(option)
 }
 
@@ -291,35 +156,6 @@ function isObject(option) {
 export function formatErrorMessage(Component, message) {
   return `${Component.moduleName}: ${message}`
 }
-
-/**
- * Schema for component config
- *
- * @typedef {object} Schema
- * @property {{ [field: string]: SchemaProperty | undefined }} properties - Schema properties
- * @property {SchemaCondition[]} [anyOf] - List of schema conditions
- */
-
-/**
- * Schema property for component config
- *
- * @typedef {object} SchemaProperty
- * @property {'string' | 'boolean' | 'number' | 'object'} type - Property type
- */
-
-/**
- * Schema condition for component config
- *
- * @typedef {object} SchemaCondition
- * @property {string[]} required - List of required config fields
- * @property {string} errorMessage - Error message when required config fields not provided
- */
-
-/**
- * @internal
- * @typedef {keyof ObjectNested} NestedKey
- * @typedef {{ [key: string]: string | boolean | number | ObjectNested | undefined }} ObjectNested
- */
 
 /* eslint-disable jsdoc/valid-types --
  * `{new(...args: any[] ): object}` is not recognised as valid
