+++ packages/govuk-frontend/src/govuk/components/notification-banner/notification-banner.mjs
@@ -1,5 +1,5 @@
-import { mergeConfigs, setFocus } from '../../common/index.mjs'
-import { normaliseDataset } from '../../common/normalise-dataset.mjs'
+import { mergeConfigs, normaliseDataset } from '../../common/configuration.mjs'
+import { setFocus } from '../../common/index.mjs'
 import { GOVUKFrontendComponent } from '../../govuk-frontend-component.mjs'
 
 /**
@@ -86,5 +86,5 @@ export class NotificationBanner extends GOVUKFrontendComponent {
  */
 
 /**
- * @typedef {import('../../common/index.mjs').Schema} Schema
+ * @typedef {import('../../common/configuration.mjs').Schema} Schema
  */
